fixed temp channel delition

+++ cogs/tempChannels.py
@@ -28,11 +28,10 @@ class TempChannels(commands.Cog):
         # Удаление временного канала
         if before.channel != None:
             if self.bot.db_cursor.execute(f"SELECT channel_id FROM temp_channels_{before.channel.guild.id} WHERE channel_id = {before.channel.id}").fetchone() != None:
-                if len(before.channel.members) == 0:
-                    if before.channel.guild.get_channel(before.channel.id):
-                        await before.channel.delete()
-                    self.bot.db_cursor.execute(f"DELETE FROM temp_channels_{before.channel.guild.id} WHERE channel_id = {before.channel.id}")
-                    self.bot.db_connection.commit()
+                if len(before.channel.members) == 0 and before.channel.guild.get_channel(before.channel.id):
+                    await before.channel.delete()
+                self.bot.db_cursor.execute(f"DELETE FROM temp_channels_{before.channel.guild.id} WHERE channel_id = {before.channel.id}")
+                self.bot.db_connection.commit()
         
 def setup(bot):
     bot.add_cog(TempChannels(bot))
